requirements.txt
Django==4.1.7
django-cors-headers==4.1.0
djangorestframework==3.14.0
mysql-client==0.0.1
mysql-connector-python==8.1.0
mysqlclient==2.2.0
requests==2.31.0
PyMySQL==1.1.0
djangorestframework-simplejwt==5.2.2
cryptography==41.0.3